air_quality_api/requirements.txt
Flask
python-dotenv
pymongo
redis
requests
Flask-Cors